requirements.txt
yapf==0.20.2
PyYAML==3.12
numpy==1.14.1
docker==3.1.1
colorama==0.3.9